REFACTOR - Langugage selector component
