+++ src/components/LanguageSelector/data.ts
@@ -1,0 +1,22 @@
+export const languages = [
+  {
+    code: 'en',
+    name: 'English',
+    flag: 'gb'
+  },
+  {
+    code: 'pt',
+    name: 'Português',
+    flag: 'br'
+  },
+  {
+    code: 'es',
+    name: 'Español',
+    flag: 'es'
+  },
+  {
+    code: 'fr',
+    name: 'Français',
+    flag: 'fr'
+  }
+]
