+++ src/components/LanguageSelector/types.ts
@@ -1,0 +1,4 @@
+export interface RenderFlagProps {
+  code: string
+  name: string
+}
